modified borrowers list columns
