+++ src/app/container/Borrower/BorrowerList.scss
