+++ src/app/index.js
@@ -9,6 +9,7 @@ import './index.css';
 import 'font-awesome/css/font-awesome.css';
 import Home from './components/Home/Home';
 import DashBoard from './components/UserPages/DashBoard/DashBoard.jsx';
+import BorrowerList from './container/Borrower/BorrowerList.jsx';
 
 render(
 <Provider store={store}>
@@ -16,7 +17,9 @@ render(
             <Route path="/" component={App} >
                 <IndexRoute component={Home} />
             </Route>
-            <Route path="/dashBoard" component={DashBoard} />
+            <Route path="/dashBoard" component={DashBoard} >
+                <IndexRoute component={BorrowerList} />                
+            </Route>
         </Router>
     </Provider>,
     window.document.getElementById('app')
